+++ packages/f2elint/src/utils/constants.ts
@@ -115,3 +115,24 @@ export const MARKDOWN_LINT_FILE_EXT: string[] = ['.md'];
  * markdownLint 扫描忽略的文件或文件目录
  */
 export const MARKDOWN_LINT_IGNORE_PATTERN: string[] = ['node_modules/', 'build/', 'dist/'];
+
+/**
+ * Prettier 扫描文件扩展名
+ */
+export const PRETTIER_FILE_EXT = [
+  ...STYLELINT_FILE_EXT,
+  ...ESLINT_FILE_EXT,
+  ...MARKDOWN_LINT_FILE_EXT,
+];
+
+/**
+ * Prettier 扫描忽略的文件或文件目录
+ */
+export const PRETTIER_IGNORE_PATTERN: string[] = [
+  'node_modules/**/*',
+  'build/**/*',
+  'dist/**/*',
+  'lib/**/*',
+  'es/**/*',
+  'coverage/**/*',
+];